fix(net): bring up the workload guest's eth0 — libkrun workloads had no egress
A libkrun workload guest got no network through ANY gateway: real gvproxy
(wide open, no policy) logged only "waiting for clients" — the guest sent no
DHCP and no egress frames, so any `nc`/connect failed. The builder VM (same
libkrun + gvproxy) works because its Rust PID-1 explicitly brings eth0 up +
runs udhcpc; the mkGuest workload `/init` brought up only loopback and relied
on libkrun's NET_FLAG_DHCP_CLIENT, which does not configure the interface here.

Fix: share one guest-side network bring-up between the builder and the workload.

- Move the guest-side network config (parse_ipv4, encode_iface_name,
  configure_static + ioctls) from mvm-build to mvm-guest::guest_net (it belongs
  in the guest crate), and add `configure_guest_network(iface, cmdline,
  fallback_ip)` = eth0 link-up (before udhcpc, busybox chicken-and-egg) → seed
  /etc/resolv.conf to the gateway resolver (ro rootfs) → udhcpc → static
  fallback on the gvproxy backends (QEMU uses its own ip= autoconfig).
- The builder's `setup_network` now delegates to it (no behavior change).
- `mvm-guest-netinit` (runs uid 0 before the agent on every workload) calls it
  before installing the mandatory-deny blackhole routes, so the workload guest
  finally gets eth0 + DHCP.
- mvm-build::guest_net keeps only the host-side console-log classifier
  (`classify_builder_net_bootstrap`), importing parse_ipv4 from mvm-guest.

Cross-compiles for aarch64-unknown-linux-musl (host-vm-init via the mvmctl
embed, mvm-guest-netinit + stage0-init via zigbuild); host tests green
(366 mvm-build + 301 mvm-guest, incl. moved guest_net + classifier tests).

Unblocks Plan 193 WS-2.2d live validation (native rvproxy enforcement can't be
observed until a libkrun workload gets network at all).
