+++ crates/mvm-build/src/bin/mvm-host-vm-init.rs
@@ -2102,7 +2102,7 @@ mod linux {
         // with EADDRNOTAVAIL. Bring it up first, independent of eth0/DHCP:
         // a network:None builder VM has no eth0 at all yet still needs
         // working loopback. Non-fatal, mirroring the eth0 bring-up below.
-        if let Err(e) = bring_iface_up("lo") {
+        if let Err(e) = mvm_guest::guest_net::bring_iface_up("lo") {
             eprintln!(
                 "mvm-host-vm-init: bring_iface_up lo failed: {e} \
                  (continuing — guest-internal loopback services \
@@ -2110,192 +2110,11 @@ mod linux {
             );
         }
 
-        // The rootfs mounts ro and /etc/resolv.conf is a baked regular file,
-        // not a /run symlink. busybox's DHCP script writes /etc/resolv.conf
-        // directly and gets EROFS, so leased DNS never lands. Mount a tmpfs-
-        // backed copy over /etc/resolv.conf before udhcpc runs so the script's
-        // write goes to tmpfs. The seed chooses the gateway resolver for the
-        // active VMM's virtual network: gvproxy backends (libkrun, Vz) serve
-        // DNS at 192.168.127.1; QEMU slirp at 10.0.2.3. Both do host-side
-        // resolution and work on any network; baked public resolvers don't.
-        // Seeding before udhcpc also means the bind-mount is in place on the
-        // Vz path where DHCP fails — the static fallback below gives the IP
-        // and the seed covers DNS.
-        // Non-fatal: a failed seed/bind-mount degrades DNS to the baked
-        // resolvers but must not block the iface-up + DHCP below — a
-        // leased-but-degraded guest beats one with no IP at all.
-        if let Err(e) = seed_resolv_conf(&cmdline) {
-            eprintln!("mvm-host-vm-init: resolv.conf seed skipped: {e} (continuing)");
-        }
-
-        // busybox 1.36.x udhcpc binds a PF_PACKET raw socket to
-        // `eth0` and `sendto`s a DHCPDISCOVER. virtio-net's eth0
-        // post-probe state is administratively DOWN, so the first
-        // sendto returns ENETDOWN and udhcpc loops forever
-        // ("broadcasting discover" → "Network is down" → reopen
-        // socket). Older udhcpc versions auto-issued
-        // SIOCSIFFLAGS|IFF_UP; modern busybox expects the caller
-        // to. The `/etc/udhcpc/default.script` hook brings the
-        // link up via `ip link set ... up`, but it only fires
-        // after udhcpc gets a lease — which requires the link
-        // already be up. Chicken-and-egg, broken by doing the
-        // ioctl ourselves before spawning udhcpc.
-        if let Err(e) = bring_iface_up("eth0") {
-            eprintln!(
-                "mvm-host-vm-init: bring_iface_up eth0 failed: {e} \
-                 (continuing — udhcpc will surface a clearer error \
-                 if the link is genuinely absent)"
-            );
-        }
-
-        // `/bin/udhcpc` — udhcpc is a busybox applet, and mkGuest
-        // installs busybox applet symlinks under `/bin/<applet>`,
-        // not `/sbin/`. (Prior `/sbin/udhcpc` hardcoding ENOENTed at
-        // every boot, which `setup_network` swallowed as non-fatal —
-        // leaving the builder VM with no DHCP-assigned IP and the
-        // inner nix build unable to reach `cache.nixos.org`.)
-        let script = "/etc/udhcpc/default.script";
-        let mut cmd = Command::new("/bin/udhcpc");
-        cmd.args(["-i", "eth0", "-n", "-q"]);
-        if std::path::Path::new(script).is_file() {
-            cmd.args(["-s", script]);
-        }
-        let status = cmd
-            .status()
-            .map_err(|e| format!("spawn /bin/udhcpc: {e}"))?;
-        if dhcp_fallback_applies(&cmdline, status.success()) {
-            // gvproxy's virtual subnet is fixed (192.168.127.0/24, gateway+DNS
-            // at .1, first DHCP client at .3). Each builder VM gets its own
-            // gvproxy instance, so the static address cannot collide.
-            eprintln!(
-                "mvm-host-vm-init: udhcpc exit {} — falling back to static \
-                 gvproxy addressing",
-                status.code().unwrap_or(-1)
-            );
-            mvm_build::guest_net::configure_static(
-                "eth0",
-                "192.168.127.3",
-                "255.255.255.0",
-                "192.168.127.1",
-            )?;
-        } else if !status.success() {
-            return Err(format!("udhcpc exit {}", status.code().unwrap_or(-1)));
-        }
-        Ok(())
-    }
-
-    /// True when a DHCP failure should trigger the static gvproxy fallback.
-    ///
-    /// The QEMU/slirp backend uses a different subnet (10.0.2.x) with its
-    /// own `ip=` autoconfig; applying the gvproxy address there would give
-    /// the wrong gateway and break connectivity. A success exit never needs
-    /// the fallback regardless of backend.
-    fn dhcp_fallback_applies(cmdline: &str, udhcpc_success: bool) -> bool {
-        if udhcpc_success {
-            return false;
-        }
-        !cmdline.split_whitespace().any(|t| t == "mvm.backend=qemu")
-    }
-
-    /// The gateway-local resolver for the active VMM's virtual network.
-    ///
-    /// QEMU user-mode networking serves DNS at 10.0.2.3; the gvproxy
-    /// backends (libkrun, Vz) at 192.168.127.1. Host-side resolution
-    /// through the gateway works on any network — baked public resolvers
-    /// only work where the local network permits direct external UDP/53.
-    fn resolver_seed(cmdline: &str) -> &'static [u8] {
-        if cmdline.split_whitespace().any(|t| t == "mvm.backend=qemu") {
-            b"nameserver 10.0.2.3\n"
-        } else {
-            b"nameserver 192.168.127.1\n"
-        }
-    }
-
-    /// Stage the gateway-resolver seed on tmpfs and bind-mount it over
-    /// the baked read-only `/etc/resolv.conf` so udhcpc's script (and
-    /// the lease it carries) can write DNS config through the mount.
-    fn seed_resolv_conf(cmdline: &str) -> Result<(), String> {
-        std::fs::create_dir_all("/run/mvm").map_err(|e| format!("mkdir /run/mvm: {e}"))?;
-        std::fs::write("/run/mvm/resolv.conf", resolver_seed(cmdline))
-            .map_err(|e| format!("seed /run/mvm/resolv.conf: {e}"))?;
-        let st = Command::new("/bin/busybox")
-            .args([
-                "mount",
-                "--bind",
-                "/run/mvm/resolv.conf",
-                "/etc/resolv.conf",
-            ])
-            .status()
-            .map_err(|e| format!("spawn mount --bind resolv.conf: {e}"))?;
-        if !st.success() {
-            return Err(format!(
-                "bind-mount resolv.conf exit {}",
-                st.code().unwrap_or(-1)
-            ));
-        }
-        Ok(())
-    }
-
-    /// Bring a network interface administratively up via
-    /// `ioctl(SIOCSIFFLAGS, IFF_UP)`. Equivalent to
-    /// `ip link set dev <iface> up`, but issued directly so we
-    /// don't pin a new path-dependency in the builder rootfs and
-    /// the error message names the failing ioctl. Called before
-    /// `udhcpc` in [`setup_network`].
-    fn bring_iface_up(iface: &str) -> Result<(), String> {
-        let name = mvm_build::guest_net::encode_iface_name(iface)?;
-
-        // SAFETY: socket(2) returns -1 on error (checked) or a
-        // valid fd. We close it on every return path below.
-        let sock = unsafe { libc::socket(libc::AF_INET, libc::SOCK_DGRAM, 0) };
-        if sock < 0 {
-            return Err(format!(
-                "socket(AF_INET, SOCK_DGRAM) for {iface}: {}",
-                std::io::Error::last_os_error()
-            ));
-        }
-
-        let result = (|| {
-            // SAFETY: `ifreq` is repr(C); zero-init + per-variant
-            // union assignment is the standard pattern. We read
-            // `ifru_flags` only after SIOCGIFFLAGS populated it.
-            let mut ifr: libc::ifreq = unsafe { std::mem::zeroed() };
-            ifr.ifr_name = name;
-
-            // libc 0.2.182's ioctl request parameter is the per-target
-            // `Ioctl` alias — `c_ulong` (u64) on linux-gnu, `c_int`
-            // (i32) on musl. Cast the SIOC* request to `libc::Ioctl` so
-            // this compiles on both the musl host-bin deployment target
-            // and a native-gnu `cargo build --all-targets` (CI).
-            if unsafe { libc::ioctl(sock, libc::SIOCGIFFLAGS as libc::Ioctl, &mut ifr) } < 0 {
-                return Err(format!(
-                    "SIOCGIFFLAGS {iface}: {}",
-                    std::io::Error::last_os_error()
-                ));
-            }
-            // SAFETY: SIOCGIFFLAGS just populated `ifru_flags`, so
-            // reading it is well-defined. Writing the OR'd value back
-            // through the same union variant is also well-defined per
-            // the Rust reference (all variants of `__c_anonymous_ifr_ifru`
-            // are `Copy`).
-            unsafe {
-                let flags = ifr.ifr_ifru.ifru_flags;
-                ifr.ifr_ifru.ifru_flags = flags | (libc::IFF_UP as libc::c_short);
-            }
-            if unsafe { libc::ioctl(sock, libc::SIOCSIFFLAGS as libc::Ioctl, &ifr) } < 0 {
-                return Err(format!(
-                    "SIOCSIFFLAGS {iface} IFF_UP: {}",
-                    std::io::Error::last_os_error()
-                ));
-            }
-            Ok(())
-        })();
-
-        // SAFETY: sock is owned by this function until close.
-        unsafe {
-            libc::close(sock);
-        }
-        result
+        // eth0 bring-up + DHCP + static fallback (.3) via the shared guest-net
+        // helper — the same path the workload guest netinit uses. The static
+        // fallback applies on the gvproxy backends only; QEMU/slirp uses its
+        // own `ip=` autoconfig.
+        mvm_guest::guest_net::configure_guest_network("eth0", &cmdline, "192.168.127.3")
     }
 
     fn run_job(cmd_sh: &str) -> (i32, String) {
@@ -2941,69 +2760,6 @@ mod linux {
             assert!(virtiofs_mount_flags("mvm-bins").contains(MsFlags::MS_RDONLY));
             assert_eq!(virtiofs_mount_flags("out"), MsFlags::empty());
             assert_eq!(virtiofs_mount_flags("job"), MsFlags::empty());
-        }
-
-        #[test]
-        fn dhcp_fallback_applies_non_qemu_failure() {
-            assert!(dhcp_fallback_applies(
-                "console=hvc0 root=/dev/vda rw",
-                false
-            ));
-        }
-
-        #[test]
-        fn dhcp_fallback_applies_qemu_no_fallback() {
-            // QEMU uses a different subnet; the static gvproxy address must not
-            // be applied there.
-            assert!(!dhcp_fallback_applies(
-                "console=hvc0 mvm.backend=qemu root=/dev/vda",
-                false
-            ));
-        }
-
-        #[test]
-        fn dhcp_fallback_applies_success_no_fallback() {
-            assert!(!dhcp_fallback_applies(
-                "console=hvc0 root=/dev/vda rw",
-                true
-            ));
-        }
-
-        #[test]
-        fn dhcp_fallback_applies_qemu_prefix_not_matched() {
-            // A token that merely starts with "mvm.backend=qemu" must not
-            // match — whole-word only.
-            assert!(dhcp_fallback_applies(
-                "console=hvc0 mvm.backend=qemux root=/dev/vda",
-                false
-            ));
-        }
-
-        #[test]
-        fn resolver_seed_gvproxy_backend() {
-            assert_eq!(
-                resolver_seed("console=hvc0 root=/dev/vda rw"),
-                b"nameserver 192.168.127.1\n"
-            );
-        }
-
-        #[test]
-        fn resolver_seed_qemu_backend() {
-            assert_eq!(
-                resolver_seed("console=hvc0 mvm.backend=qemu root=/dev/vda"),
-                b"nameserver 10.0.2.3\n"
-            );
-        }
-
-        #[test]
-        fn resolver_seed_qemu_token_must_be_whole_word() {
-            // A token that starts with "mvm.backend=qemu" but has trailing
-            // characters is not a match — the seed must stay at gvproxy's
-            // gateway resolver.
-            assert_eq!(
-                resolver_seed("console=hvc0 mvm.backend=qemux root=/dev/vda"),
-                b"nameserver 192.168.127.1\n"
-            );
         }
 
         /// `run_job_streaming` calls the
